+++ exercises/matthew_cumber/ex02_oo_basics/src/Shape.h
@@ -11,8 +11,8 @@ public:
   virtual ~Shape() = default;
 
   /// methods to be implemented in derived classes
-  virtual float area() = 0;
-  virtual float perimeter() = 0;
+  virtual double area() = 0;
+  virtual double perimeter() = 0;
 
   /// Getters for all shapes
   inline std::string getType() const { return type; }
